node: move seed node implementation to its own file
